Updated with Estimate V1.1
Fixed the error causing it to crash. Bathrooms being an integer would sometimes make it crash when there would be a half bath, updated to convert to float instead. 

Planning to edit this to be just one function with a conditional statement about additional cleaning services requested instead and just combine into one class with keyword parameters for the function incase extra services are not requested.